Moved ArtistList into more generic GenericList
