+++ frontend/src/components/ArtistCard/ArtistCard.js
