+++ frontend/src/components/ArtistCard/ArtistCard.scss
