+++ frontend/src/components/GenericList/GenericList.scss
